Fix rubberbanding when playing high fps and/or ping.
The original implementation of `GetFreeMove()` limits the saved moved
history to 100 points. On high fps and/or high ping, it can easily
exceed over 100. This causes "rubberbanding" as the client is replaying
incomplete move history after it receives an authoritative position
update from the server.

Unfortunately, `GetFreeMove()` is marked final in the PlayerController
class. To workaround that, `GetFreeMoveEx()` is introduced with a
configurable `MaxSavedMoves` value. Everything is encapsulated in the
`ModernPlayer` class, which is intended to be subclassed instead of
`xPlayer`.